Adjust the old brcc for its new home in brcc/v01.
Make the !WIN32 code path build correctly since the Makefile's hardcoding
gcc, even on Windows.

Add -Wall and eliminate the resulting compiler warnings.

NULL terminate in makestring.pl to avoid string overruns and append a \n
while we're at it to silence a compiler warning.

Mess with the logging in codegen.c and teach it to tolerate cgc output with
^M's in it.  It remains to be seen whether I'll need to resort to filtering
out all the trailing ^M's, but for now we'll try to interoperate.