Allow commandline selection of backends.  By default brcc will compile for the CPU and ps20 backends.  Instead, a single backend may be specified with
	brcc -p [ ps20 | fp30 | arb | cpu ]

(arb doesn't do anything, but v01 had a #define for it).  It's only possible
to specify a single backend with -p for now, but I think that's likely to
meet the needs of most people who would want it.

While I was at it, I adjusted common.mk to pass $(BRCCFLAGS) when compiling
.br files so you can test with

	make BRCCFLAGS='-p ps20'

etc. And I think that between us, Ian and I have figured out how to force
make to preserve the intermediate .cpp files when compiling a .br.

I also updated the brcc 00-TODO a little.