Another wave of iterator syntax adjustments.  Change from
	iter var<100> = BRTIter(1.0, 10.0);

to

	iter var<100> = iter(1.0, 10.0);

This seemingly minor change required torquing the grammar since 'iter' is a
reserved word (whereas it'd just smile and nod at BRTIter), but this brought
the added reward that the grammar now enforces some more of the syntax at
parse time.

Testing the grammar error handling demonstrated that it really didn't know
how to recover from parse errors in float constructors or iterators, so I
taught it how to do that too, which should reduce nicely the amount of
cascading errors.

And I updated all of prog/ to use the new proper iter syntax.