(1) Code factoring.  There's a lot less duplicated code.  I could pass streams as arguments, there'd be even less.
(2) As a nice side effect, it was really easy to regularize the printing
code.  The output of readwrite and runkernel should both be fairly easy to
read.

(3) I added a lot of intermediate lengths to readwrite so I could see the
curve better.  I took out the high run lengths from runkernel because they
take a long time on my card and at 1024x1024 the nv30gl runtime starts
crashing / returning garbage.

(4) Darn precision.  I taught the mismatch code to tolerate up to a 10%
error between output and expected output because numbers beyond about 17
million are fuzzy on the nv30 and about 131 thousand on ATI.