(1) Turn back on the mismatch checking and fix a bug in 10% error tolerance checking.
(2) Write a simple native C version of the HardWork kernel so I could test
against something more representative than the CPU runtime.

(3) Jgraph scripts and makefile alterations to generate graphs for each of
the backends.  Jgraph isn't the most flexible language in the world, so
there's a lot of copy pasting, but it's a start.