Took out the toplevel BRCCFLAGS.  They were making it impossible to override the flags per-Makefile or easily on the command line.
Fix common.mk so that errors during the regression test abort the whole
make, not just the nested one.

Removed all the regression tests that don't currently pass from the list of
tests.

BRCC: Made 'getQualifiers()' a const method and fixed up all the subclasses.

BRCC: Converted 'kernel' from a storage class (which it wasn't) to a type
qualifier (which it mostly is).  It's now just like 'reduce'.

BRCC: Type checking for kernel calls.  There are three checks:
	- Correct number of arguments
	- Correct stream-ness / non-stream-ness of arguments
	- Compatible base type of arguments

Unfortunately, kernel-kernel calls violate both of the first two rules
(internal indexof calls apparently require extra arguments and at kernel
time, whether a parameter is a stream or not is very fluid).  Long term the
right thing to do is make the type checker smart enough to know whether a
callsite is inside a kernel or not, but for now brcc takes a -N flag to type
check with the looser rules for nested calls.

Extra unfortunately, the 'syntax' of scatterOp is impossible to type-check
effectively.  The regression test looks like:

	reduce void mygets(float a<>, float b<>, float c) { ... }
	...
	streamScatterOp(a, b, e, mygets(1));

That 'mygets(1)' syntax is a complete monkey-wrench to any sort of normal
type checking approach.  For now, brcc takes a -t flag that disables all
type-checking and I've added it to the scatter test's Makefile.

Extra extra unfortunately, some of the tests circumvent all type checking by
hiding some C++ only code in a header file, including it, and then passing
structs to kernels and relying upon the C++ compilation phase to magically
(via operator overloading) convert the structs back to streams.  This
totally violates any hope of type checking.  I've added -t to the Makefile
of all the tests that currently do this.

Fixed up all of the tests that were violating the type checking rules.