Add support for passing stream arguments to non-kernel functions along with an example in the accumulate regression test.
While I was there, I did some class hierarchy surgery.  BrtStreamType and
BrtIterType looked remarkably similar and I disliked the weird naming that
would have come from BrtStreamType and BrtStreamParamType.  So now
BrtStreamType is the abstract parent and there are 'initialized',
'iterator', and 'parameter' children.

All the regression tests pass with both the CPU and nv30gl runtimes as of
this checkin.