Cope with streams whose type is a typedef that's just a base type.  I.e.
	typedef float4 streamType;

	kernel k(streamType a<>, out streamType b<>);

I did it the arguably dumb way of rewriting the type in codegen.  The
obvious alternative is to teach BaseType::getBase() to actually dig out the
real base type when it's a typedef of some sort, but that's a more subtle
change than I felt like making to the compiler after this long apart from
it.